tg: support $tgdisplayac for --abort/--continue display
The $tgdisplay variable shows the correct tg command to the user
to re-run the same copy of tg with the same -C, -u and -r options
and is used in various places.

Introduce a new $tgdisplayac for use with --abort/--continue displays
that should not show any -u or -r options because they are implicit
in the --continue operation (and irrelevant for --abort).

Update the status display and tg-update.sh to use the new $tgdisplayac
variable.  Also simplify the maintenance of $tgdisplay at the same time
and correct a latent flaw where "-u" was inadvertently omitted due to
an unfortunate typo.  While doing this, the display used when multiple
-C options are present has been somewhat improved -- previously it
would only show the last -C option (which was wrong) and now it will
still only show one -C option but it will be the effective $PWD if
more than one -C option was given.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>