tg: replace $tg with tg() and export TG_EXPLICIT_REMOTE
Several tg commands recursively invoke tg for various
functions.  They have been doing this like so:

    $tg subcomand arguments

This does not work if the path to "tg" contains spaces.
The reason it's been left unquoted is so that any
additional -r or -u options can be passed in to the
subcommand to make sure it operates properly.

Instead introduce TG_EXPLICIT_REMOTE and export it
whenever the -r or -u option is used.  Then pick it up
for the default value of $base_remote on startup.

This eliminates the need to pass any -r or -u options
along when recursively invoking tg.

The -C option(s) effect will be passed along implicitly
via the current working directory and any -c options
via the exported GIT_CONFIG_PARAMETERS.  This eliminates
the need to pass any options when recursiely invoking tg.

But $tg still might not be an absolute path.  That should
be okay as it would only be left that way if it were found
in the $PATH, but just to be safe, set it to an absolute
path.  Also change its name from $tg to $tgbin to help
catch stragglers.

At this point then those recursive invocations could become:

    "$tgbin" subcommand arguments

But that's rather ugly.  Instead define a helper function
named "tg" that does exactly that and then the recursive
invocations can simply become like this:

    tg subcommand arguments

Very much nicer, easier to remember and less error prone.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>