tg-update.sh: give progress messages some structure
It's nearly impossible to follow along in the status output
without pausing the display or reviewing it later.

Attempt to improve on this by splitting the recursive path
onto its own line with a distinctive "==> " prefix.

While this hardly results in crystal clarity, it does make
it somewhat easier to parse quickly.

Also stop setting PS1 at the same time as that hasn't been
necessary since tg update dropped the subshell scheme.

Furthermore, make sure that when showing the continuation
instructions for a manual merge that any prefix is
eliminated entirely.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>