tg-update.sh: remove special case and allow auto rerere continue
The git_topmerge utility function had one special case where
it would attempt to use the normal "git merge" for --merge
mode (which is only used when merging a remote branch [not
it's base, the branch itself] into the local branch).

This mode could only be triggered when the remote branch
had unmerged updates but that would mean they'd been fetched
from somewhere.  As a result, it's not a mode that triggers
very often.  It will only trigger if new changes have
appeared in the remote branch and tg update is run with
remote updates enabled and the changes cannot be picked up
via trivial merge.

In other words, it doesn't get much exercise.

Instead, just make it use the same logic that runs
merge-recursive directly.  It works just as well (with a few
minor tweaks) for the --merge mode as it does for the other
modes and very minor efficiency savings by using "git merge"
in this one special case just don't add up to make it worthwhile
since the case is triggered so seldomly.  This nicely eliminates
a dark corner hiding place for bugs too.

Now that all the merge modes are going through the same
merge-recursive driver logic, move up the call to git rerere
to just before checking to see if the merge left any unresolved
files or not.

This is the big win.

If rerere.enabled is true AND rerere.autoupdate is true then
the call to "git rerere" will not only resolve a previously
resolved conflict, it will also stage it.  With the normal
"git merge" that would not be enough to keep going because it
happens too late.

However, with the new merge-recursive driver logic it now
happens before the check to see if the merge is complete
which means that when both those rerere settings are enabled
the merge will AUTOMATICALLY CONTINUE if rerere resolves and
stages a resolution!

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>