tg-update.sh: --aggressive optimizations
An earlier change stopped using --aggressive with read-tree -m
when in "ours" and "theirs" modes so that the index-merge-one-file
helper would get a chance to do the right thing in the cases
--aggressive additionally handles but which might not end up
actually matching "ours" or "theirs".

However, by swiping some of the same code git_topmerge uses, that's
no longer necessary and use of --aggressive can be restored for
all modes when doing index-only read-tree -m merges.  Make it so.
This also has the side benefit of enforcing "ours" and "theirs"
even when read-tree -m without --aggressive would have done the
wrong thing.  Another win.

Mostly this change is neutral unless there are a lot of the kinds
of changes that --aggressive resolves happening on a regular basis,
in which case it's a big win.

The other optimization eliminates an unnecessary extra call to
git write-tree.  While not a huge win it does net almost a whopping
2% speed increase.  Nothing earth shaking, but nothing to sneeze
at either.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>