tg-create.sh: make sure creation does not lose .topdeps or .topmsg
The tg create command necessarily writes new .topdeps and .topmsg
files.  The overwrite checking was suboptimal when an unborn
branch was involved.

Correct this by using the new ensure_clean_topfiles function to
try and guarantee that if .topdeps and/or .topmsg are present
in the working tree they are also exist in an identical version
in the index and HEAD commit.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>