tg-create.sh: prevent existing "clean" .topfiles from aborting creation
Make sure that any pre-existing .topdeps and/or .topmsg files are blown
away (provided they are "clean" of course) before attempting to write
the new ones or add them to the index.

Also make sure that the new branch starts out with a "clean" slate.

This prevents unexpected errors from occurring and is primarily of a
preventative nature, but it doesn't hurt to be careful.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>