tg.sh: minor branch_needs_update efficiency
The recurse_deps function has been providing the
_dep_annihilated and _dep_has_remote variables for
a bit now.  Use them in the branch_needs_update
function to save time instead of recomputing them.

While this is a nice efficiency, once the values
are computed they are cached which means that while
this is a nice and correct update it doesn't really
net very much in the way of performance gain.

It's a measurable speed improvement, but only barely.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>