tg-summary.sh: awksome accelerate --list --verbosely
Use the run_awk_topgit_branches and run_awk_topgit_msg
functions to extremely accelerate the "--list" function
of tg summary.

As a bonus, if the new "--verbose" (or just "-v") option
is also used then include the subjects in the list.

It's now blazingly fast with or without --verbose
whereas before it was okay but nowhere near fast enough
to include the subject lines too.

Also sneak in support for spelling out "-t" as "--terse".

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>