tg-patch.sh: get all awked up on --from
The `git am` command really blows chunks when handed a patch
that has no From: header line in it.

Avoid this by having `tg patch` warn about such patches by
default (if the .topmsg file does not have a From: line in
it, tg patch does not normally add one by default).

Provide new --from[=<addr>], --no-from and --quiet options
to insert a missing From: header with the specified address
or defaulting to the value of the last Signed-of-by: line
in the .topmsg body or finally the current GIT_AUTHOR_IDENT.

The new config option `topgit.from` can be a boolean to
enable or disable the --from option or a non-boolean value
to enable it with a specific address.  The special value
"quiet" enables the --quiet option to stop complaining about
such patches and leave their From: lines alone.

The default is, of course, just to warn and leave the .topmsg
alone.  (In no case is the actual .topmsg file modified, just
the resulting patch output.)

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>