tg-info.sh: introduce --series[=<head>] option
Using the new awking awesome acceleration, make the --series
option quickly show the list of all patches that are part of
the patch series that contains the given branch (or HEAD by
default), in the order they would be linearized as well as
navigated over and with an indicator mark to quickly locate
the specified branch if it's not the last one in the list.

The output provides a quick and simple means to find the
"YOU ARE HERE" mark and reorient oneself after perhaps
spending some time away engaging in distracting activities.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>