tg: awksome accelerate list_deps by banishing it
The list_deps function was just painfully slow.  The only
client was tg-summary.sh.  Replace the two calls to list_deps
in tg-summary.sh with something much speedier and then
abolish the list_deps function from tg.sh.

Two new helper functions are added to tg.sh to assist with
properly setting up the -td= and/or -tm= option values in order
to imitate the effect of the "-i" (index) and "-w" (worktree)
options.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>