tg-checkout: awking awesome --series of overhauls
Completely and totally (well, there might still be the
odd few lines that match the original) overhaul tg-checkout.sh

Now using the awksome functions makes it wickedly fast.

It's also now possible to step by something other than 1
or "--all" or start from somewhere other than "HEAD'.

For the "goto" mode the same "--series[=<head>]" option
results in a list (remarkably like `tg info --series`) of
branches in the current (or specified) series to choose from.

When movement is ambiguous and a list is shown, the containing
patch series is/are shown to assist with the selection process.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>