tg-annihilate.sh: do not use `tg prev` or `tg next`
In the future the semantics of tg prev and tg next will
match those of the "next" and "prev" options to "tg checkout".

At that point tg annihilate would break.

Avoid this by switching to using `tg summary --deps` instead.

Save the output to a temporary file and then it can easily
be used for both functions.  Grabbing it before actually
annihilating the branch is more robust which is a bonus and
eliminates need for the hacky "-w" option.

Some other minor improvements snuck in to make sure the
.topfiles are clean and the branch has not already been
annihilated.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>