tg-info.sh: support --deps and --dependents options
The "--deps" option shows a list of non-annihilated TopGit
branches that are dependencies of the current (or specified)
TopGit branch (i.e. the ".topdeps" contents but with non-TopGit
and annihilated branches filtered out).

The "--dependents" option shows a list of non-annihilated TopGit
branches that depend on the current (or specified) TopGit branch.

The semantics of `tg next` and `tg prev` will change in the
future to regain sanity and parity with `tg checkout`.

With this change no functionality (whatever it might have been
or have been perceived to be) will be lost when "tg next" and
"tg prev" get overhauled; it will still be available somewhere
(depending on what the perceived or desired output of the old
"tg prev" and/or "tg next" was in the mind of the user).

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>