tg: previous to next overhaul
Completely overhaul tg-next.sh and tg-prev.sh.

They now navigate exactly as expected and can move
in increments of other than one step at a time.

They can also show the containing series name(s) of
the result(s) with --verbose (aka -v).

They are the "show, but don't go" versions of the
"tg checkout" command's "next" and "previous" options.

The "tg info" command provides the ability to view
a specific branch's immediate dependencies or
dependents if desired, but that is not the purpose
of the "tg next" and "tg prev" commands.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>