tg-update.sh: automatically set rerere.autoUpdate by default
Now that TopGit drives the merge machinery itself during update
merges, if "rerere" automatically stages resolutions for all
conflicts, the update will automatically just keep going.

Temporarily set rerere.autoUpdate to true while tg update is
running by default to enable this when rerere has been enabled.

This will not enable the "rerere" feature, merely make it auto-
stage if it has been enabled.  This is a huge time saver and
while it may not always be the desired behavior in general,
while running "tg update" it almost always is.

Of course there are two new options --no-auto/--no-auto-update
and --auto/--auto-update to explicitly control this as well as
a new "topgit.setAutoUpdate" config option that when explicitly
set to false will change TopGit's default to leaving the
"rerere.autoUpdate" setting unmolested.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>