t/helper/recurse_deps_internal: more correctly parse args
The first argument to recurse_deps_internal is not necessarily
a branch name.

However, it's most helpful to have that check.

Rather than removing the check, iterate over the arguments
and require that at least one of them is a valid TopGit branch
name with the bonus of transforming 'HEAD' or '@' into the
actual name for easier testing.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>