Makefile: add "target" target
Add a special "target" target that always runs Makefile.mak and,
if the variable "TARGET" has been set, passes those target(s)
along as the goals.

Normally this is not necessary, but when making explicit targets
that are usually made implicitly this may be necessary in obscure
cases to force them to build properly rather than having the wrapper
Makefile incorrectly think they're already up-to-date.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>