Makefile.mak: compute TG-BUILD-SETTINGS differently
Instead of trying to define a multi-line value that's legible,
works across all POSIX makes and is insensitive to make whitespace
handling, define BUILD_SETTINGS to be a shell function definition.

This way the produced result looks much nicer and it's impervious
to the vagaries of various POSIX makes' whitespace handling.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>