awk: make sure close function calls use parentheses
This is not POSIX compatible awk code:

    filename = "test.txt"
    print "hello" >filename
    close filename

However, some awk implementations silently accept it without
complaint.  POSIX specifies the "close" function to be called
using parentheses and the POSIX-compatible version is thus this:

    filename = "test.txt"
    print "hello" >filename
    close(filename)

This problem is a regression introduced in topgit-0.19.7.

Add the missing parentheses in both places they were omitted in
order to restore correct operation on all systems.

Thanks-to: Paul Molodowitch <pm@stanfordalumni.org>
Signed-off-by: Kyle J. McKay <mackyle@gmail.com>