t/test-lib-main.sh: include $this_test in comments
Instead of outputting something like this:

  # passed all 77 test(s)

Include the $this_test name so it looks like this instead:

  # t0001 passed all 77 test(s)

Do the same for all similar comments.  Update the test-lib
testing test (t0001) to grok the new format.

When using a parallel job capable make to run the "test" test
target in multiple job mode having $this_test in the output
comments makes the result so much more intelligible.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>