README: add a "TESTING TOPGIT" section
Explain how to run the tests in various ways and point at the
documentation for the testing library at the same time.

Tweak the requirements section a bit to link to the new section
and mention the need for perl and fix a few minor formatting
issues in there as well.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>