t/Makefile.mak: introduce TESTLIB_MAKE_OPTS make variable
Pass $(TESTLIB_MAKE_OPTS) to the sub-make that runs the individual tests
when not using prove (i.e. the t/Makefile.mak "test" target aka the default
and what DEFAULT_TEST_TARGET=test uses).

This allows, for example, using:

  make TESTLIB_MAKE_OPTS="-j 4" test

to run multiple individual tests in parallel (when not using the "prove"
target) and the "make" in question fails to properly pass down a "-j 4"
option given directly to the top-level make command.

Add some documentation updates to point out this new option.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>