test-lib: move BASH_XTRACEFD assignment into test_lib_main_init_specific
When the test library was being adapted to do caching and avoid
duplication of effort when running more than one test script, the
assignment to BASH_XTRACEFD was inadvertently moved into the
test_lib_main_init_generic function.

This turns out not to matter unless test caching is taking place
and the shell being used as the POSIX shell is actually version 4
or later of bash.

Fix this by moving the BASH_XTRACEFD assignment into the
test_lib_main_init_specific function where it belongs and stop
caching its value as restoring the cached value (which always
happens to be 4) will cause an error if FD 4 has not been set up
yet and the running shell happens to be bash version 4 or later.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>