test-lib-main.sh: workaround broken bash
Newer versions of bash do not handle this correctly:

  bash -c 'fail_() { return $1; }; (eval "set -e; fail_ 5"); echo $?'

The correct output is 5.  Newer versions of bash output 1 (older
versions of bash get it right).

Turning off "set -e" mode before leaving the eval works around the
problem (as does using an explicit "exit" but an explicit "exit" is
undesirable for other reasons).

Enhance the fail_ function so it can be used to disable "set -e"
mode before returning and use that to workaround the problem.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>