test-lib-main.sh: enable line numbers when using bash
Getting line numbers into the output requires use of
command aliases.  All the other shells except bash always
have them enabled.

If the "shopt" command is supported, attempt to enable
alias expansion using the necessary bashism so that line
numbers will show up properly in the output.

Cache the result of the check for the shopt command when
caching is enabled.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>