test-lib-functions.sh: allow test_commit to skip the tag
If the <tag> name that would be used is the empty string or contains
any whitespace, simply skip the tagging step.

This is more flexible and provides a means to add commits without
constantly adding new tags too.

Also correct the emission of <message> to use `printf '%s\n'` instead
of echo to make sure nothing untoward happens to the message.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>