all: unset ksh errors
If the shell is ksh then this:

  unset it && unset it && echo good || echo bad

Outputs "bad" instead of "good".

POSIX clearly stipulates that using unset on a variable
that has never be set is NOT an error and MUST NOT cause
a non-zero status.

Tolerate this brokennes by using an internal 'unset_' function
that ignores any unset error status or the 'sane_unset' function
in test scripts or by adding an explicit '|| :' where neither of
those may be expected to be available.

Add a t0002-sanity test to document this (and possibly other future)
"tolerate_failure" issues.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>