all: unset and unalias zsh traps
zsh is NOT a POSIX conforming shell!

However, it's close; but it has a number of maddening
defects that make it nearly unusable as such.

Nevertheless, attempt to make it work anyway provided
it's invoked under a suitable name in order to activate
its (poor quality) "sh" emulation mode.

The t/sanity test serves as documentation for most of
the various nasties that have to be accomodated.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>