test-lib-main: bless use of $color for debug output
When producing debugging output with, for example,
the test_debug function, it's nice to see the output
colorized, but only when color should be used.

In the test environment Git will never automatically
activate color due to the way the test environment
has been set up.  Configuring Git to automatically
enable color would therefore be of no help.  Configuring
Git to always output color when "color" is active for
the test enviornment could break scripts that are
testing Git's output.

Ergo the color variable.  Since it's already being
set up to indicate whether or not color output is active
in the test environment, bless its use by scripts by
documenting it in the README-TESTLIB file and changing
the true value to be "--color" so that it can easily be
used as an argument to git commands when color output
is desired in debugging output.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>