t/t0003-testlib-git.sh: add tests for git utility functions
While the "testlib-basic" test covers most of the rest of the
functions, it does not cover Git-specific functions.

Since this is not a test suite for Git itself, tests at that
level are not appropriate.

However, several of the provided utility functions are
Git-centric so provide some simple tests to make sure they
provide the functionality they're supposed to provide.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>