testlib: allow "test_plan'?'"
To document that a plan line has been deliberately omitted from
the beginning of the TAP output (perhaps the test in question
always produces a variable number of subtests), an explicit:

    test_plan \?

line can be used.  (The "?" must be quoted in some fashion to
avoid wildcard expansion.  Although, the setup for individual tests
does not start out with any single character file names in the
directory so it should still work without quoting in case it's
been accidentally omitted.)

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>