test-lib-functions.sh: make test_expect_code work with set -e
In contrast to Git's version of the test library, this version
runs all subtests in a subshell by default with "set -e" active.

Fix the "test_expect_code" function to work properly in a "set -e"
environment so that when the command in question returns a non-zero
code the "test_expect_code" function does not prematurely abort.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>