tg: refactor remote top-bases sniff out code
Move the remote top-bases sniff out code from tg-remote.sh to
a new function, check_remote_topbases, in tg.sh.

This way tg-remote.sh gets simpler, the check is reusable and
the error messages for that check are kept in the same file near
the error messages for the normal top-bases check so they can be
kept sychronized if they ever need to be changed.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>