tg.sh: allow --top-bases to work for remotes
Make "--top-bases -r" show the correct top-bases location for
the current TopGit remote.

Also show help for the option with -h or --help and accept a
remote name after the "-r" to override any otherwise set remote.

Use the same logic that "tg remote" uses to sniff out the correct
top-bases location for the remote.

Add tests for the new functionality.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>