tg: document and tolerate "bare branches"
Document what a "bare branch" is (a TopGit branch with neither a
.topdeps nor .topmsg file) and tolerate them even more than they
are already tolerated by allowing continued commits on them without
complaints from the pre-commit hook.

Support for this type of TopGit branch has been present for some
time.  This just formalizes it, documents it and removes the need
for use of "-n" when committing on such a branch.

Update the tests for them as well to match the new pre-commit
hook behavior.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>