tg: call ensure_work_tree where needed
Some commands work in bare repositories.  Others do not.

Call ensure_work_tree for the commands (or parts thereof) that
actually do require a work tree to be present to function properly.

This avoids beginning the action only to have it fail later
when a Git command ends up failing and having it possibly write
work tree types of files into a bare repository.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>