tg: find added lines missing final \n
If the last line of a file does not end in a newline, git diff
will include this precious gem:

    \ No newline at end of file

Accommodate that and do not drop out of being "in_hunk" when looking
for added lines.

Unfortunately while that does cure the problem of missing added lines
because we dropped out of "in_hunk" mode, the line with the missing
trailing \n will end up being seen as a newly added line since it
appears both in the "-" and "+" section as a result of now having a
\n at the end of it.

To solve that problem, run git diff with the --ignore-space-at-eol option
so that we get the diffs we are expecting (which kind of makes the fix
for the '\ No newline at end of file' moot, but it does need to be fixed
anyway).

It's extremely non-trivial to attempt to "fix up" a diff that contains
the '\ No newline at end of file' items into one that doesn't so that
there are no spurious "new" lines.  However, it's better to possibly
find a spurious "new" line than miss a whole bunch of other real "new"
lines, so the fix is worth keeping in case future code uses diff_added_lines
on something and either forgets to apply the --ignore-space-at-eol option
or it's not feasible to do so (perhaps the diff was produced externally).

Add a new set of tests to check this.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>