tg-export.sh: minor correction and efficiency
Replace big old unnecessarily long pipeline being used to resolve
a tag object into the commit it points to (if it dereferences to a
commit rather than something else) with a single git rev-parse call.

There is some question whether anything but a commit object can
even end up in such a location in the first place anymore.

However, since there are no tests for tg-export (yet) keep the
prior semantics intact with the replacement.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>