aggregate-results.sh: purge expr and show all failures
There's no reason to shell out to expr just to extract a prefix
of the filename.

Purge use of expr in favor of simple parameter expansion stripping
options.

At the same time make sure no failed tests are omitted from the
list of failed test names just because they don't match the
pattern /^t\d+-/.  Instead include the entire name in the list
(except for the .counts suffix) in that case.

Normally only test names matching the pattern would ever be passed
in, but during debugging test scripts may have different names and
they need to be included in the failure list if they fail rather
than just being included in the failure count and leaving one to
wonder which tests actually failed.

Of course prove could be used instead but it doesn't really support
full test verbose mode therefore it's good to fix this problem and
get rid of another gratuitous use of expr in the process.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>