tg-create: require format.signoff for active sob line
Since topgit-0.19, tg create by default runs an editor on
the initial version of the `.topmsg` file for a new branch
(unless suppressed with `--no-edit`).

When that change was made, a "Signed-off-by:" line started being
unconditionally added to the initial template.

Unfortunately the old behavior was also retained in that if
Git's `format.signoff` config is true a "Signed-off-by:" line
is added.  Which means that as of topgit-0.19 if `format.signoff`
is true you get two copies of the "Signed-off-by:" line in the
initial `.topmsg` template file and neither is commented out.

Correct this travesty.

Always emit exactly one "Signed-off-by:" line at the end of the
template.  But, if the `format.signoff` config is NOT set to a
true value then also comment it out with a leading '#' so it will
automatically be stripped if no action is taken.

This restores the semantics of the `format.signoff` config to the
proper behavior (thus repairing the regression) while at the same
time making it easy to add a "Signed-off-by:" line without making
any typographical errors by simply stripping the leading '#' if
that's desired and `format.signoff` is not already set to true.

(And since `tg` accepts `-c name=value` options it's also possible
to temporarily alter the `format.signoff` value for a single use
of `tg create`.)

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>