tg-info.sh: show a message when the remote base is missing
Instead of just quietly skipping the remote base check when the
remote base is missing, emit a message mentioning that it is missing.

The user might not realize the base(s) are missing and emitting a
message will help clarify the issue.  When top-bases are in the old
location they will not automatically be fetched by default unless the
`tg remote` command has been run first (yet another reason to switch
to the new top-bases location).

The remote heads themselves, though, will almost always be present
allowing the presence of a remote head to be detected so that the
absense of a remote base can be diagnosed.

Many commands will just silently ignore such an orphaned remote head.

Therefore it's good to have at least one command that will show
something informative instead and `tg info` seems like a good choice.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>