testlib: pimp out test_plan
Strongly encourage use of the `test_plan` function to output a TAP
test plan line in advance rather than relying on `test_done` to do it.

Emit a request to do so if `test_plan` was not called before `test_done`
(unless the entire test was skipped courtesy of TESTLIB_SKIP_TESTS).

Also use all the arguments after the first to form the skip message with
a plan count of 0 instead of just the second one.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>