gnomake: suppress unwanted gratuitous "Entering directory ..." lines
If the "Entering directory ..." lines have not been explicitly
requested via the -w (aka --print-directory) option then suppress
them for all but real sub make operations.

But do it carefully so it works with makes that do not understand
the option.

Also take this opportunity to sneak in two new gnomake functions
vplus and vpre.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>