t0001: add export unset works test
POSIX is very explicit about this.

The "export" command works on unset variables.  There's even a special
format for `export -p` output for unset variables.

It's very particular about specifying that marking variables for export
"shall cause them to be in the environment of subsequently executed commands."

Since the environment is just an array of "name=value" strings and the "="
does not appear to be optional, an unset but exported variable that appears
"in the environment of subsequently executed commands" will therefore be
indistinguishable from an exported variable of the same name set to the
empty string.

Nevertheless, POSIX hath spoken, make it so and add a test for this
unspeakable behavior.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>